refactor: added use of constants
