--- src/constants/index.ts
@@ -1,3 +1,0 @@
-// Central export for all constants
-export { COLORS, type ColorType } from './colors';
-export { COOKING_POSITIONS, type CookingPositionType } from './positions';
