+++ src/hooks/useImageModal.ts
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import type { ColorType, ModalState } from '../types';
-import { COLORS } from '../constants';
+import { COLORS } from '../constants/colors';
 
-const DEFAULT_IMAGE: ColorType = COLORS.RED; // fallback predeterminado
+const DEFAULT_IMAGE: ColorType = COLORS.RED;
 
 export const useImageModal = () => {
   const [modalState, setModalState] = useState<ModalState>({
